add test for `PsWorkflow`
