+++ src/dolphin/workflows/config/__init__.py
@@ -2,4 +2,5 @@
 
 from ._common import *  # noqa
 from ._displacement import *  # noqa
+from ._ps import *  # noqa
 from ._enums import *  # noqa
